+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.5.4] - 2026-04-07
+- UIRoute now remembers the last selection until closed
+- UIRouter now uses private properties and exposes a getter for the current stack
+
 ## [0.5.3] - 2026-04-07
 - Fix to Transition driver to call OnTransitionStart/Complete used by some sample transitions like the Text one
 - Add WindowResolutionManager to set window stuff
 - Added random option to LoadingScreenTextObject.
+- Changed Scene controller to use unscaledDeltaTime in case the game is paused
 
 ## [0.5.2] - 2026-04-07
 - Added/Fixed Vector3/Int Extension stuff